14/03/2022 -- Prueba instalacion. Vista menu producto 2

+++ modulo1_mss/models/models.py
@@ -27,7 +27,7 @@ class producto(models.Model):
     refProducto = fields.Char(string='Referencia', required=True)
     marca = fields.Char(string='Marca', required='True')
     tipoProducto = fields.Selection(string='Tipo de producto', selection=[('a', 'Electrodomestico'),('b', 'Telefono'),('c', 'Ordenador portátil'),('d', 'Consola')], help='Tipo del producto.')
-    precioCompra = fields.Double(string='Precio compra', required=True)
+    precioCompra = fields.Decimal(string='Precio compra', required=True)
     precioVenta = fields.Double(string='Precio venta', required=True)
 
     proveedor_id = fields.Many2one('modulo1_mss.proveedor', string='Proveedor')
